+++ boss-bes-core/boss-bes-core-data/src/main/java/com/boss/bes/core/data/pojo/basevo/BaseInfoVO.java
@@ -28,11 +28,11 @@ public abstract class BaseInfoVO<T extends BaseModVO> {
     /**
      * 当前查询条件下总共的数据条数
      */
-    private Integer dataCount;
+    private Long dataCount;
 
     public BaseInfoVO() {}
 
-    public BaseInfoVO(List<T> dataList, Integer pageSize, Integer pageNum, Integer dataCount) {
+    public BaseInfoVO(List<T> dataList, Integer pageSize, Integer pageNum, Long dataCount) {
         this.dataList = dataList;
         this.pageSize = pageSize;
         this.pageNum = pageNum;
@@ -63,11 +63,11 @@ public abstract class BaseInfoVO<T extends BaseModVO> {
         this.pageNum = pageNum;
     }
 
-    public Integer getDataCount() {
+    public Long getDataCount() {
         return dataCount;
     }
 
-    public void setDataCount(Integer dataCount) {
+    public void setDataCount(Long dataCount) {
         this.dataCount = dataCount;
     }
